created a new class container
