+++ src/components/ClassContainer/class_container.scss
@@ -1,0 +1,4 @@
+.classesDiv {
+  display: flex;
+  margin-bottom: 2em;
+}
