+++ src/components/Home/MainContainer/index.js
@@ -3,10 +3,8 @@
 // ----  { Libraries } ----
 import React, { useState, useContext, useEffect } from "react";
 import { createStyles, makeStyles } from "@material-ui/core/styles";
-import { Link } from "react-router-dom";
 
 // ----  { Routes, ActionTypes etc. Custom variables. } ----
-import * as ROUTES from "../../../constants/routes";
 
 // ----  { Styles } ----
 import "./main_container.scss";
@@ -17,36 +15,19 @@ import firebase from "../../Firebase";
 // ----  { Render Components } -----
 import Typography from "@material-ui/core/Typography";
 import Container from "@material-ui/core/Container";
-import Button from "@material-ui/core/Button";
-import Icon from "@material-ui/core/Icon";
-import Fab from "@material-ui/core/Fab";
 import { Context } from "../../../context";
 import Loading from "../../Loading";
+import Icon from "@material-ui/core/Icon";
+import Fab from "@material-ui/core/Fab";
 import CreateClassModal from "../CreateClassModal";
 import EditClassModal from "../EditClassModal";
-
-const useStyles = makeStyles(theme =>
-  createStyles({
-    fab: {
-      margin: theme.spacing(1),
-      width: "40px",
-      height: "40px"
-    },
-    add: {
-      position: "absolute",
-      bottom: "0"
-    },
-    extendedIcon: {
-      marginRight: theme.spacing(1)
-    }
-  })
-);
+import ClassContainer from "../../ClassContainer";
 
 export default function MainContainer(props) {
   const {
     myClasses: { loading, myClasses }
   } = props;
-  const classes = useStyles();
+
   const [createClassModalState, setCreateModalState] = useState(false);
   const [editClassModalState, setEditClassModalState] = useState({
     showModal: false,
@@ -111,37 +92,13 @@ export default function MainContainer(props) {
           {loading ? (
             <Loading text="Hämtar in dina klasser"></Loading>
           ) : myClasses ? (
-            myClasses.map(schoolClass => (
-              <div className="classesDiv" key={"div " + schoolClass.id}>
-                <Button
-                  variant="contained"
-                  color="primary"
-                  classes={{ root: "classesButton", label: "classesLabel" }}
-                  to={ROUTES.HOME}
-                  component={Link}
-                  size="medium"
-                  key={"button class " + schoolClass.id}
-                >
-                  {schoolClass.className}
-                </Button>
-                <Fab
-                  color="secondary"
-                  aria-label="edit"
-                  className={classes.fab}
-                  key={"fab " + schoolClass.id}
-                  onClick={() => editThisClass(schoolClass.id)}
-                >
-                  <Icon>edit</Icon>
-                </Fab>
-                <Fab
-                  key={"delete button " + schoolClass.id}
-                  aria-label="delete"
-                  className={classes.fab}
-                  onClick={() => deleteFromDB(schoolClass.id)}
-                >
-                  <Icon>delete</Icon>
-                </Fab>
-              </div>
+            myClasses.map(({ className, id }) => (
+              <ClassContainer
+                className={className}
+                id={id}
+                editThisClass={editThisClass}
+                deleteFromDB={deleteFromDB}
+              ></ClassContainer>
             ))
           ) : (
             <div>Du har inga klasser</div>
