+++ src/components/Home/MainContainer/main_container.scss
@@ -1,7 +1,3 @@
-.classesDiv {
-  display: flex;
-  margin-bottom: 2em;
-}
 .classesButton {
   flex: 1;
 }
